refactor mask_to_roi, add tests (still failing)
